[pickman] Merge patch series "fs: exfat: Fix some Smatch issues"

See merge request u-boot/u-boot!625

+++ tools/fit_image.c
@@ -826,7 +826,7 @@ static int fit_handle_file(struct imgtool *itl)
 	char tmpfile[MKIMAGE_MAX_TMPFILE_LEN];
 	char bakfile[MKIMAGE_MAX_TMPFILE_LEN + 4] = {0};
 	char cmd[MKIMAGE_MAX_DTC_CMDLINE_LEN];
-	size_t size_inc;
+	int size_inc;
 	int ret = EXIT_FAILURE;
 
 	/* Flattened Image Tree (FIT) format  handling */
